refactor(landingpage): extract contributers section to own component
