+++ packages/landingpage-new/components/about/contributors.module.scss
@@ -1,0 +1,19 @@
+@use '../../styles/typography.scss';
+
+h1.header {
+  @include typography.font($size: 50px, $weight: lighter);
+  color: #232527;
+  letter-spacing: 0.53px;
+  margin-top: 0;
+  margin-bottom: 20px;
+
+  b {
+    font-weight: bold;
+  }
+}
+
+.container {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 25px;
+}
