--- packages/landingpage-new/pages/about/about.module.scss
@@ -1,15 +1,0 @@
-@use '../../styles/typography.scss';
-
-h1.header {
-  @include typography.font($size: 50px, $weight: lighter);
-  color: #232527;
-  letter-spacing: 0.53px;
-  margin-top: 0;
-  margin-bottom: 20px;
-}
-
-.container {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 25px;
-}
